requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pymodbus==3.6.9
pyserial==3.5
python-periphery==2.4.1

influxdb-client==1.45.0
